Fixed fielding precision in Teams.csv.
Updated the FP column in Teams.csv, fixing the truncation issue for yearids
1871-1999. These fielding percentages came from an earlier Baseball
Databank release's "Teams.txt" file, contained in the following ZIP file's
URL:

http://www.baseball-databank.org/files/BDB-csv-2011-03-28.zip

In every case for the affected seasons, the 2011 file's fielding
percentages tenths and hundredths "places" matched those in the current
file. The only difference is that the 2011 version had a thousandths
place.

This should address issue #9.