Re-sort Appearances table.
This changes the sort order of the Appearances table.  Instead of
sorting by club within a season, then person ID, the sort is by person ID
first, then by club.  This is probably more intuitive, and now
matches the sort order generated by the tool that updates this file.
In this way, the diffs on future commits will reflect only actual
changes rather than changes in canonical ordering of rows.